Старт переделки в последовательную обработку потоковых данных
Также исключено задание переменных через консоль (теперь только в main.cpp менять).
Вместо обработки определенного кол-ва файлов видео, будут последовательно обрабатываться
выхваченный кадр с первым кадром (не окончательно доделано).
После обработки смещение обработанного кадра относительно первого записываеся в .txt файл
(название файла указывается также в main.cpp).
При вычислениях считается, что первый кадр имеет центр (0,0) и угол поворота = 0.

+++ frameprocessor.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 using namespace cv;
 
-FrameProcessor::FrameProcessor(std::string detectorName, float threshold, int planes)
+FrameProcessor::FrameProcessor(std::string detectorName, int threshold, int planes)
 	: m_threshold(threshold)
 {
     m_pDetector = BRISK::create(threshold, planes);
@@ -34,7 +34,7 @@ ImageData FrameProcessor::MatchImages(cv::InputArray& im1, FeatureInfo & first, 
 	return ret;
 }
 
-float FrameProcessor::SetThreshold(float newThreshold)
+int FrameProcessor::SetThreshold(int newThreshold)
 {
 	cv::Ptr<cv::BRISK> ptr = m_pDetector.dynamicCast<cv::BRISK>();
 	ptr->setThreshold(newThreshold);
